fix: 消除全部 Kotlin 编译警告并修复空指针隐患
- 修复 SafeDns hostAddress 可空调用与公告时间解析 null 隐患
- 图标迁移至 AutoMirrored（ArrowBack/Send/InsertDriveFile/VolumeOff/KeyboardArrowRight）
- ColorScheme 构造补全 fixed 容器角色；menuAnchor 迁移新重载
- 移除废弃的 statusBarColor/navigationBarColor 赋值（enableEdgeToEdge 已处理）
- Locale 改用 forLanguageTag；Room 迁移带参 fallbackToDestructiveMigrationOnDowngrade
- GlobalScope 替换为应用级 CoroutineScope（复用 AnimeTrackApp.appScope）
- 补充 FlowPreview OptIn 与 @param:StringRes use-site target

+++ app/src/main/java/com/aiexile/animetrack/util/ShareCardRenderer.kt
@@ -174,7 +174,7 @@ object ShareCardRenderer {
         }
 
         // 右侧信息 - 标题（最多2行）
-        val titleLayout = StaticLayout.Builder.obtain(anime.title, 0, anime.title.length, titlePaint, infoWidth.toInt())
+        val titleLayout = StaticLayout.Builder.obtain(anime.title, 0, anime.title.length, titlePaint, infoWidth)
             .setMaxLines(2)
             .setEllipsize(TextUtils.TruncateAt.END)
             .build()
